// src/services/models/memeResponse.ts
/**
 * Generated by orval v7.1.0 🍺
 * Do not edit manually.
 * A8 Pump Fun API
 * OpenAPI spec version: 0.0.12
 */
import type { MemeResponseStatus } from "./memeResponseStatus";
import type { UserResponse } from "./userResponse";

export interface MemeResponse {
  readonly created_at: string;
  readonly createdTime: number;
  readonly creator: string;
  readonly decimals: string;
  readonly description: string;
  readonly dexScreenerPairUrl: string;
  readonly discord: string;
  readonly factory: string;
  readonly id: number;
  readonly image: string;
  readonly isWhiteList: boolean;
  readonly kingOfTheHillProgress: string;
  readonly kingOfTheHillTime: number;
  readonly liquidity: string;
  readonly listProgress: string;
  readonly marketCap: string;
  readonly name: string;
  readonly nativeOffset: string;
  readonly pairAddress: string;
  readonly price24hChange: string;
  readonly priceByNative: string;
  readonly priceByUsd: string;
  readonly pumpAddress: string;
  readonly saleAmount: string;
  readonly status: MemeResponseStatus;
  readonly symbol: string;
  readonly telegram: string;
  readonly tokenAddress: string;
  readonly tokenOffset: string;
  readonly totalSupply: string;
  readonly updated_at: string;
  readonly userInfo: UserResponse;
  readonly virtualNativeReserve: string;
  readonly virtualTokenReserve: string;
  readonly volume24hs: string;
  readonly volumeBuyByNative: string;
  readonly website: string;
  readonly whiteListEndTime: number;
  readonly whiteListStartTime: number;
  readonly x: string;
}